@@ -1,15 +1,12 @@
-[tool.poetry]
+[project]
 name = "common"
 version = "0.1.0"
-description = ""
-authors = ["Your Name <you@example.com>"]
+description = "Shared models and utilities for microservices and AI agents"
 readme = "README.md"
-
-[tool.poetry.dependencies]
-python = "^3.10"
-pydantic = "^2.8.2"
-
-
-[build-system]
-requires = ["poetry-core"]
-build-backend = "poetry.core.masonry.api"
+requires-python = ">=3.12"
+dependencies = [
+    "pydantic>=2.9.2",
+    "httpx>=0.27.0",
+    "langchain>=0.3.0",
+    "langchain-openai>=0.2.0",
+]
