@@ -7,6 +7,8 @@ import os
 from datetime import datetime
 from typing import Any
 
+from common import MCPGrafanaClient
+
 logger = logging.getLogger(__name__)
 
 
@@ -18,10 +20,7 @@ class LogsAnalyzer:
     def __init__(self):
         """Initialize logs analyzer with MCP client"""
         self.mcp_url = os.getenv("MCP_GRAFANA_URL", "http://grafana-mcp:8000")
-        
-        # Import MCP client (will be available from agents-common)
-        # For now, we'll use a mock implementation
-        self.mcp_client = None
+        self.mcp_client = MCPGrafanaClient(base_url=self.mcp_url)
 
     async def close(self):
         """Close MCP client"""
@@ -36,28 +35,28 @@ class LogsAnalyzer:
     ) -> dict[str, Any]:
         """
         Analyze logs based on user query
-        
+
         Args:
             query: User query
             time_range: Time range for analysis
             context: Additional context (services, focus, etc.)
-            
+
         Returns:
             Analysis results
         """
         context = context or {}
         logger.info(f"Analyzing logs for query: {query}")
-        
+
         # Build LogQL query
         logql_query = self._build_logql_query(query, context)
         logger.info(f"Generated LogQL: {logql_query}")
-        
+
         # Query Loki via MCP (mock for now)
         logs_data = await self._query_loki(logql_query, time_range)
-        
+
         # Analyze log data
         analysis = self._analyze_logs(logs_data, query, context)
-        
+
         return {
             "agent_name": "logs",
             "analysis": analysis["summary"],
@@ -70,73 +69,63 @@ class LogsAnalyzer:
     def _build_logql_query(self, query: str, context: dict[str, Any]) -> str:
         """
         Build LogQL query from user query and context
-        
+
         Args:
             query: User query
             context: Context with services, focus, etc.
-            
+
         Returns:
             LogQL query string
         """
         # Extract services from context
         services = context.get("services", [])
-        
+
         # Base query
         if services:
             service_filter = "|".join(services)
             logql = f'{{service_name=~"{service_filter}"}}'
         else:
             logql = '{job=~".+"}'
-        
+
         # Add error filter if needed
         query_lower = query.lower()
         if "error" in query_lower or "fail" in query_lower or "problem" in query_lower:
             logql += ' |= "error"'
-        
+
         # Add specific patterns
         if "db" in query_lower or "database" in query_lower:
             logql += ' |~ "(?i)(db|database).*error"'
-        
+
         return logql
 
     async def _query_loki(self, logql: str, time_range: str) -> dict[str, Any]:
         """
         Query Loki via MCP Grafana
-        
+
         Args:
             logql: LogQL query
             time_range: Time range
-            
+
         Returns:
             Logs data
         """
-        # TODO: Use actual MCP client when available
-        # For now, return mock data
         logger.info(f"Querying Loki with: {logql}")
-        
-        return {
-            "logs": [
-                {
-                    "timestamp": "2025-11-08T10:15:23Z",
-                    "service_name": "order",
-                    "level": "ERROR",
-                    "message": "Simulated DB insertion error",
-                },
-                {
-                    "timestamp": "2025-11-08T10:16:45Z",
-                    "service_name": "order",
-                    "level": "ERROR",
-                    "message": "Simulated DB insertion error",
-                },
-                {
-                    "timestamp": "2025-11-08T10:18:12Z",
-                    "service_name": "order",
-                    "level": "ERROR",
-                    "message": "Unexpected error during order creation",
-                },
-            ],
-            "total_count": 47,
-        }
+
+        # Query via MCP client
+        result = await self.mcp_client.query_logs(
+            query=logql, time_range=time_range, limit=100
+        )
+
+        # Return in expected format
+        if "error" in result:
+            logger.warning(f"Loki query error: {result['error']}")
+            # Return mock data as fallback
+            return {
+                "logs": [],
+                "total_count": 0,
+            }
+
+        return result
 
     def _analyze_logs(
         self,
@@ -146,57 +135,61 @@ class LogsAnalyzer:
     ) -> dict[str, Any]:
         """
         Analyze log data and generate insights
-        
+
         Args:
             logs_data: Raw logs from Loki
             query: Original user query
             context: Request context
-            
+
         Returns:
             Analysis with summary, data, and confidence
         """
         logs = logs_data.get("logs", [])
         total_count = logs_data.get("total_count", len(logs))
-        
+
         # Count errors by type
         error_types = {}
         affected_services = set()
-        
+
         for log in logs:
             error_msg = log.get("message", "")
             service = log.get("service_name", "unknown")
-            
+
             # Group by error type
             if error_msg in error_types:
                 error_types[error_msg] += 1
             else:
                 error_types[error_msg] = 1
-            
+
             affected_services.add(service)
-        
+
         # Build summary
         error_types_list = [
             {"type": error_type, "count": count}
-            for error_type, count in sorted(error_types.items(), key=lambda x: x[1], reverse=True)
+            for error_type, count in sorted(
+                error_types.items(), key=lambda x: x[1], reverse=True
+            )
         ]
-        
+
         summary_parts = [
             f"Found {total_count} error logs in the last {context.get('time_range', '1h')}."
         ]
-        
+
         if error_types_list:
             primary_error = error_types_list[0]
             summary_parts.append(
                 f"Primary error: '{primary_error['type']}' ({primary_error['count']} occurrences)."
             )
-        
+
         if "Simulated DB" in str(error_types):
             summary_parts.append(
                 "Note: These appear to be simulated errors (ERROR_RATE environment variable)."
             )
-        
-        summary_parts.append(f"Affected services: {', '.join(sorted(affected_services))}.")
-        
+
+        summary_parts.append(
+            f"Affected services: {', '.join(sorted(affected_services))}."
+        )
+
         return {
             "summary": " ".join(summary_parts),
             "data": {
@@ -212,11 +205,11 @@ class LogsAnalyzer:
     def _generate_grafana_links(self, logql: str, time_range: str) -> list[str]:
         """
         Generate Grafana Explore links
-        
+
         Args:
             logql: LogQL query
             time_range: Time range
-            
+
         Returns:
             List of Grafana URLs
         """
@@ -227,9 +220,8 @@ class LogsAnalyzer:
     async def check_mcp_health(self) -> bool:
         """
         Check if MCP Grafana server is reachable
-        
+
         Returns:
             True if healthy, False otherwise
         """
-        # TODO: Implement actual health check
-        return True
+        return await self.mcp_client.health_check()
