@@ -49,7 +49,7 @@ class HealthResponse(BaseModel):
 async def analyze(request: AnalyzeRequest):
     """
     Analyze observability issues based on user query
-    
+
     Routes the query to specialized agents and synthesizes their responses.
     """
     try:
@@ -69,14 +69,14 @@ async def analyze(request: AnalyzeRequest):
 async def health():
     """
     Health check endpoint
-    
+
     Checks connectivity to all specialized agents.
     """
     agent_status = await orchestrator.check_agents_health()
-    
+
     all_healthy = all(status == "reachable" for status in agent_status.values())
     status = "healthy" if all_healthy else "degraded"
-    
+
     return HealthResponse(
         status=status,
         agents=agent_status,
@@ -88,9 +88,15 @@ async def health():
 async def startup():
     """Application startup"""
     logger.info("Orchestrator Agent starting up...")
-    logger.info(f"Logs Agent URL: {os.getenv('AGENT_LOGS_URL', 'http://agent-logs:8002')}")
-    logger.info(f"Metrics Agent URL: {os.getenv('AGENT_METRICS_URL', 'http://agent-metrics:8003')}")
-    logger.info(f"Traces Agent URL: {os.getenv('AGENT_TRACES_URL', 'http://agent-traces:8004')}")
+    logger.info(
+        f"Logs Agent URL: {os.getenv('AGENT_LOGS_URL', 'http://agent-logs:8002')}"
+    )
+    logger.info(
+        f"Metrics Agent URL: {os.getenv('AGENT_METRICS_URL', 'http://agent-metrics:8003')}"
+    )
+    logger.info(
+        f"Traces Agent URL: {os.getenv('AGENT_TRACES_URL', 'http://agent-traces:8004')}"
+    )
 
 
 @app.on_event("shutdown")
