@@ -7,6 +7,8 @@ import os
 from datetime import datetime
 from typing import Any
 
+from common import MCPGrafanaClient
+
 logger = logging.getLogger(__name__)
 
 
@@ -18,10 +20,7 @@ class TracesAnalyzer:
     def __init__(self):
         """Initialize traces analyzer with MCP client"""
         self.mcp_url = os.getenv("MCP_GRAFANA_URL", "http://grafana-mcp:8000")
-        
-        # Import MCP client (will be available from agents-common)
-        # For now, we'll use a mock implementation
-        self.mcp_client = None
+        self.mcp_client = MCPGrafanaClient(base_url=self.mcp_url)
 
     async def close(self):
         """Close MCP client"""
@@ -36,28 +35,28 @@ class TracesAnalyzer:
     ) -> dict[str, Any]:
         """
         Analyze distributed traces based on user query
-        
+
         Args:
             query: User query
             time_range: Time range for analysis
             context: Additional context (services, focus, etc.)
-            
+
         Returns:
             Analysis results
         """
         context = context or {}
         logger.info(f"Analyzing traces for query: {query}")
-        
+
         # Build TraceQL query
         traceql_query = self._build_traceql_query(query, context)
         logger.info(f"Generated TraceQL: {traceql_query}")
-        
+
         # Query Tempo via MCP (mock for now)
         traces_data = await self._query_tempo(traceql_query, time_range)
-        
+
         # Analyze trace data
         analysis = self._analyze_traces(traces_data, query, context)
-        
+
         return {
             "agent_name": "traces",
             "analysis": analysis["summary"],
@@ -70,83 +69,68 @@ class TracesAnalyzer:
     def _build_traceql_query(self, query: str, context: dict[str, Any]) -> str:
         """
         Build TraceQL query from user query and context
-        
+
         Args:
             query: User query
             context: Context with services, focus, etc.
-            
+
         Returns:
             TraceQL query string
         """
         filters = []
         services = context.get("services", [])
         query_lower = query.lower()
-        
+
         # Service filter
         if services:
-            service_filter = '|'.join(services)
+            service_filter = "|".join(services)
             filters.append(f'service.name=~"{service_filter}"')
-        
+
         # Error filter
         if "error" in query_lower or "fail" in query_lower:
             filters.append("status=error")
-        
+
         # Slow/latency filter
         if "slow" in query_lower or "latency" in query_lower:
             filters.append("duration > 500ms")
-        
+
         # Build final query
         if filters:
             traceql = "{" + " && ".join(filters) + "}"
         else:
             traceql = '{service.name=~".+"}'
-        
+
         return traceql
 
     async def _query_tempo(self, traceql: str, time_range: str) -> dict[str, Any]:
         """
         Query Tempo via MCP Grafana
-        
+
         Args:
             traceql: TraceQL query
             time_range: Time range
-            
+
         Returns:
             Traces data
         """
-        # TODO: Use actual MCP client when available
-        # For now, return mock data
         logger.info(f"Querying Tempo with: {traceql}")
-        
-        return {
-            "traces": [
-                {
-                    "trace_id": "abc123",
-                    "duration_ms": 312,
-                    "status": "error",
-                    "services": ["customer", "ordercheck", "order"],
-                    "spans": [
-                        {"service": "customer", "operation": "send_order", "duration_ms": 85},
-                        {"service": "ordercheck", "operation": "validate_order", "duration_ms": 120},
-                        {"service": "order", "operation": "POST /orders", "duration_ms": 312, "status": "error"},
-                    ],
-                },
-                {
-                    "trace_id": "def456",
-                    "duration_ms": 285,
-                    "status": "ok",
-                    "services": ["customer", "ordercheck", "order", "postgres"],
-                    "spans": [
-                        {"service": "customer", "operation": "send_order", "duration_ms": 78},
-                        {"service": "ordercheck", "operation": "validate_order", "duration_ms": 95},
-                        {"service": "order", "operation": "POST /orders", "duration_ms": 267},
-                        {"service": "postgres", "operation": "INSERT", "duration_ms": 145},
-                    ],
-                },
-            ],
-            "total_count": 150,
-            "failed_count": 15,
-        }
+
+        # Query via MCP client
+        result = await self.mcp_client.query_traces(
+            query=traceql, time_range=time_range, limit=50
+        )
+
+        # Return in expected format
+        if "error" in result:
+            logger.warning(f"Tempo query error: {result['error']}")
+            # Return mock data as fallback
+            return {
+                "traces": [],
+                "total_count": 0,
+                "failed_count": 0,
+            }
+
+        return result
 
     def _analyze_traces(
         self,
@@ -156,80 +140,86 @@ class TracesAnalyzer:
     ) -> dict[str, Any]:
         """
         Analyze trace data and generate insights
-        
+
         Args:
             traces_data: Raw traces from Tempo
             query: Original user query
             context: Request context
-            
+
         Returns:
             Analysis with summary, data, and confidence
         """
         traces = traces_data.get("traces", [])
         total_count = traces_data.get("total_count", len(traces))
         failed_count = traces_data.get("failed_count", 0)
-        
+
         # Analyze durations
         durations = [t.get("duration_ms", 0) for t in traces]
         avg_duration = sum(durations) / len(durations) if durations else 0
         slow_traces = [t for t in traces if t.get("duration_ms", 0) > 300]
-        
+
         # Find bottlenecks (slowest operations)
         bottlenecks = {}
         all_services = set()
-        
+
         for trace in traces:
             for span in trace.get("spans", []):
                 service = span.get("service", "unknown")
                 operation = span.get("operation", "unknown")
                 duration = span.get("duration_ms", 0)
-                
+
                 key = f"{service}::{operation}"
                 if key not in bottlenecks:
-                    bottlenecks[key] = {"service": service, "operation": operation, "durations": []}
+                    bottlenecks[key] = {
+                        "service": service,
+                        "operation": operation,
+                        "durations": [],
+                    }
                 bottlenecks[key]["durations"].append(duration)
                 all_services.add(service)
-        
+
         # Calculate average durations for bottlenecks
         bottleneck_list = []
         for key, data in bottlenecks.items():
             avg_dur = sum(data["durations"]) / len(data["durations"])
             if avg_dur > 100:  # Only include slow operations
-                bottleneck_list.append({
-                    "service": data["service"],
-                    "operation": data["operation"],
-                    "avg_duration_ms": int(avg_dur),
-                    "count": len(data["durations"]),
-                })
-        
+                bottleneck_list.append(
+                    {
+                        "service": data["service"],
+                        "operation": data["operation"],
+                        "avg_duration_ms": int(avg_dur),
+                        "count": len(data["durations"]),
+                    }
+                )
+
         bottleneck_list.sort(key=lambda x: x["avg_duration_ms"], reverse=True)
-        
+
         # Build summary
         summary_parts = [
             f"Analyzed {total_count} traces over the last {context.get('time_range', '1h')}."
         ]
-        
+
         if failed_count > 0:
             summary_parts.append(
                 f"Found {failed_count} failed traces ({failed_count/total_count*100:.1f}% failure rate)."
             )
-        
+
         summary_parts.append(f"Average trace duration: {int(avg_duration)}ms.")
-        
+
         if slow_traces:
-            summary_parts.append(
-                f"Detected {len(slow_traces)} slow traces (>300ms)."
-            )
-        
+            summary_parts.append(f"Detected {len(slow_traces)} slow traces (>300ms).")
+
         if bottleneck_list:
             top_bottleneck = bottleneck_list[0]
             summary_parts.append(
                 f"Primary bottleneck: {top_bottleneck['service']}.{top_bottleneck['operation']} "
                 f"({top_bottleneck['avg_duration_ms']}ms average)."
             )
-        
-        summary_parts.append(f"Service dependency chain: {' → '.join(sorted(all_services))}.")
-        
+
+        summary_parts.append(
+            f"Service dependency chain: {' → '.join(sorted(all_services))}."
+        )
+
         return {
             "summary": " ".join(summary_parts),
             "data": {
@@ -246,11 +236,11 @@ class TracesAnalyzer:
     def _generate_grafana_links(self, traceql: str, time_range: str) -> list[str]:
         """
         Generate Grafana Explore links
-        
+
         Args:
             traceql: TraceQL query
             time_range: Time range
-            
+
         Returns:
             List of Grafana URLs
         """
@@ -261,9 +251,8 @@ class TracesAnalyzer:
     async def check_mcp_health(self) -> bool:
         """
         Check if MCP Grafana server is reachable
-        
+
         Returns:
             True if healthy, False otherwise
         """
-        # TODO: Implement actual health check
-        return True
+        return await self.mcp_client.health_check()
