@@ -14,18 +14,18 @@ def get_llm(
 ) -> ChatOpenAI:
     """
     Get configured LLM instance for agents
-    
+
     Uses the local Docker AI Model Runner endpoint by default.
     Can be overridden with environment variables for production use.
-    
+
     Args:
         model: Model name
         temperature: Sampling temperature (0-1)
         max_tokens: Maximum tokens in response
-        
+
     Returns:
         Configured ChatOpenAI instance
-        
+
     Environment Variables:
         LLM_BASE_URL: Base URL for LLM API (default: http://172.17.0.1:12434/engines/llama.cpp/v1)
         LLM_API_KEY: API key (default: dummy-token for local model)
